Prepare for Release 1.0

+++ app/src/main/java/org/gdg/frisbee/android/app/App.java
@@ -144,24 +144,26 @@ public class App extends Application implements LocationListener {
 
     public void migrate(int oldVersion, int newVersion) {
 
-        mPreferences.edit().remove(Const.SETTINGS_GCM_REG_ID).apply();
+        if(oldVersion < 15 || Const.ALPHA) {
+            mPreferences.edit().remove(Const.SETTINGS_GCM_REG_ID).apply();
 
-        mPreferences.edit().clear().apply();
-        mPreferences.edit().putBoolean(Const.SETTINGS_FIRST_START, true);
-        String rootDir = null;
-        if (Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState())) {
-            // SD-card available
-            rootDir = Environment.getExternalStorageDirectory().getAbsolutePath()
-                    + "/Android/data/" + getPackageName() + "/cache";
-        } else {
-            File internalCacheDir = getCacheDir();
-            rootDir = internalCacheDir.getAbsolutePath();
+            mPreferences.edit().clear().apply();
+            mPreferences.edit().putBoolean(Const.SETTINGS_FIRST_START, true);
+            String rootDir = null;
+            if (Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState())) {
+                // SD-card available
+                rootDir = Environment.getExternalStorageDirectory().getAbsolutePath()
+                        + "/Android/data/" + getPackageName() + "/cache";
+            } else {
+                File internalCacheDir = getCacheDir();
+                rootDir = internalCacheDir.getAbsolutePath();
+            }
+            deleteDirectory(new File(rootDir));
+
+            Toast.makeText(getApplicationContext(), "Alpha version always resets Preferences on update.", Toast.LENGTH_LONG).show();
+
+            mPreferences.edit().putInt(Const.SETTINGS_VERSION_CODE, newVersion).apply();
         }
-        deleteDirectory(new File(rootDir));
-
-        Toast.makeText(getApplicationContext(), "Alpha version always resets Preferences on update.", Toast.LENGTH_LONG).show();
-
-        mPreferences.edit().putInt(Const.SETTINGS_VERSION_CODE, newVersion).apply();
     }
 
     public void updateLastLocation() {
